관통프로젝트/02-pjt/requirements.txt
Django==5.2.0
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.60.1
mysqlclient==2.2.0
